Revert "refactor(ui): extract ContentViewHeader and StatusMessageView to UI/Components\n\n- Pure moves; preserve behavior and accessibility\n- Add compatibility initializers in WizardTypes to keep wizard core tests compiling\n- Optional<AutoFixAction> helper for older test inference"
This reverts commit 3f098cd58688ae2d6a344cdb2582fe38230a8542.
