--- Sources/KeyPath/UI/Components/StatusMessageView.swift
@@ -1,86 +1,0 @@
-import SwiftUI
-
-struct StatusMessageView: View {
-    let message: String
-    let isVisible: Bool
-
-    var body: some View {
-        Group {
-            if isVisible {
-                HStack(spacing: 12) {
-                    Image(systemName: iconName)
-                        .font(.title2)
-                        .foregroundColor(iconColor)
-
-                    Text(message)
-                        .font(.headline)
-                        .foregroundColor(.primary)
-
-                    Spacer()
-                }
-                .padding(.horizontal, 16)
-                .padding(.vertical, 12)
-                .background(
-                    RoundedRectangle(cornerRadius: 12)
-                        .fill(backgroundColor)
-                        .overlay(
-                            RoundedRectangle(cornerRadius: 12)
-                                .stroke(borderColor, lineWidth: 1)
-                        )
-                )
-                .shadow(color: Color.black.opacity(0.05), radius: 2, x: 0, y: 1)
-                .transition(
-                    .asymmetric(
-                        insertion: .move(edge: .bottom).combined(with: .opacity),
-                        removal: .move(edge: .bottom).combined(with: .opacity)
-                    ))
-            }
-        }
-        .animation(.spring(response: 0.5, dampingFraction: 0.8), value: isVisible)
-    }
-
-    private var iconName: String {
-        if message.contains("❌") || message.contains("Error") || message.contains("Failed") {
-            "xmark.circle.fill"
-        } else if message.contains("⚠️") || message.contains("Config repaired")
-            || message.contains("backed up") {
-            "exclamationmark.triangle.fill"
-        } else {
-            "checkmark.circle.fill"
-        }
-    }
-
-    private var iconColor: Color {
-        if message.contains("❌") || message.contains("Error") || message.contains("Failed") {
-            .red
-        } else if message.contains("⚠️") || message.contains("Config repaired")
-            || message.contains("backed up") {
-            .orange
-        } else {
-            .green
-        }
-    }
-
-    private var backgroundColor: Color {
-        if message.contains("❌") || message.contains("Error") || message.contains("Failed") {
-            Color.red.opacity(0.1)
-        } else if message.contains("⚠️") || message.contains("Config repaired")
-            || message.contains("backed up") {
-            Color.orange.opacity(0.1)
-        } else {
-            Color.green.opacity(0.1)
-        }
-    }
-
-    private var borderColor: Color {
-        if message.contains("❌") || message.contains("Error") || message.contains("Failed") {
-            Color.red.opacity(0.3)
-        } else if message.contains("⚠️") || message.contains("Config repaired")
-            || message.contains("backed up") {
-            Color.orange.opacity(0.3)
-        } else {
-            Color.green.opacity(0.3)
-        }
-    }
-}
-
