+++ Sources/KeyPath/UI/ContentView.swift
@@ -667,7 +667,54 @@ struct ContentView: View {
     }
 }
 
-// ContentViewHeader moved to UI/Components/ContentViewHeader.swift
+struct ContentViewHeader: View {
+    @ObservedObject var validator: MainAppStateController // 🎯 Phase 3: New controller
+    @Binding var showingInstallationWizard: Bool
+    @EnvironmentObject var kanataManager: KanataViewModel // Phase 4: MVVM
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 2) {
+                HStack(alignment: .firstTextBaseline, spacing: 8) {
+                    Button(action: {
+                        AppLogger.shared.log(
+                            "🔧 [ContentViewHeader] Keyboard icon tapped - launching installation wizard")
+                        showingInstallationWizard = true
+                    }, label: {
+                        Image(systemName: "keyboard")
+                            .font(.title2)
+                            .foregroundColor(.blue)
+                    })
+                .buttonStyle(PlainButtonStyle())
+                .accessibilityIdentifier("launch-installation-wizard-button")
+                .accessibilityLabel("Launch Installation Wizard")
+                .accessibilityHint("Click to open the KeyPath installation and setup wizard")
+
+                Text("KeyPath")
+                    .font(.largeTitle.weight(.bold))
+
+                Spacer()
+
+                // System Status Indicator in top-right
+                SystemStatusIndicator(
+                    validator: validator,
+                    showingWizard: $showingInstallationWizard,
+                    onClick: {
+                        kanataManager.requestWizardPresentation()
+                    }
+                )
+            }
+
+                Text("Record keyboard shortcuts and create custom key mappings")
+                    .font(.subheadline)
+                    .foregroundColor(.secondary)
+                    .padding(.top, 0)
+        }
+        .frame(maxWidth: .infinity, alignment: .leading)
+        .padding(.horizontal, 12)
+        .padding(.vertical, 10)
+        // Transparent background - no glass header
+    }
+}
 
 // RecordingSection moved to UI/Components/RecordingSection.swift
 
@@ -804,7 +851,89 @@ struct DiagnosticSummarySection: View {
     }
 }
 
-// StatusMessageView moved to UI/Components/StatusMessageView.swift
+struct StatusMessageView: View {
+    let message: String
+    let isVisible: Bool
+
+    var body: some View {
+        Group {
+            if isVisible {
+                HStack(spacing: 12) {
+                    Image(systemName: iconName)
+                        .font(.title2)
+                        .foregroundColor(iconColor)
+
+                    Text(message)
+                        .font(.headline)
+                        .foregroundColor(.primary)
+
+                    Spacer()
+                }
+                .padding(.horizontal, 16)
+                .padding(.vertical, 12)
+                .background(
+                    RoundedRectangle(cornerRadius: 12)
+                        .fill(backgroundColor)
+                        .overlay(
+                            RoundedRectangle(cornerRadius: 12)
+                                .stroke(borderColor, lineWidth: 1)
+                        )
+                )
+                .shadow(color: Color.black.opacity(0.05), radius: 2, x: 0, y: 1)
+                .transition(
+                    .asymmetric(
+                        insertion: .move(edge: .bottom).combined(with: .opacity),
+                        removal: .move(edge: .bottom).combined(with: .opacity)
+                    ))
+            }
+        }
+        .animation(.spring(response: 0.5, dampingFraction: 0.8), value: isVisible)
+    }
+
+    private var iconName: String {
+        if message.contains("❌") || message.contains("Error") || message.contains("Failed") {
+            "xmark.circle.fill"
+        } else if message.contains("⚠️") || message.contains("Config repaired")
+            || message.contains("backed up") {
+            "exclamationmark.triangle.fill"
+        } else {
+            "checkmark.circle.fill"
+        }
+    }
+
+    private var iconColor: Color {
+        if message.contains("❌") || message.contains("Error") || message.contains("Failed") {
+            .red
+        } else if message.contains("⚠️") || message.contains("Config repaired")
+            || message.contains("backed up") {
+            .orange
+        } else {
+            .green
+        }
+    }
+
+    private var backgroundColor: Color {
+        if message.contains("❌") || message.contains("Error") || message.contains("Failed") {
+            Color.red.opacity(0.1)
+        } else if message.contains("⚠️") || message.contains("Config repaired")
+            || message.contains("backed up") {
+            Color.orange.opacity(0.1)
+        } else {
+            Color.green.opacity(0.1)
+        }
+    }
+
+    private var borderColor: Color {
+        if message.contains("❌") || message.contains("Error") || message.contains("Failed") {
+            Color.red.opacity(0.3)
+        } else if message.contains("⚠️") || message.contains("Config repaired")
+            || message.contains("backed up") {
+            Color.orange.opacity(0.3)
+        } else {
+            Color.green.opacity(0.3)
+        }
+    }
+}
 
 struct DiagnosticSummaryView: View {
     let criticalIssues: [KanataDiagnostic]
